+++ R/generate_cre_dataset.R
@@ -6,10 +6,12 @@
 #'  Causal Rule Ensemble package
 #'
 #' @param n the number of observations
-#' @param rho the correlation within the variables
-#' @param n_rules the number of causal rules, either 2 or 4
-#' @param effect_size the effect size magnitude
-#' @param binary whether to use binary or continuous outcomes, either TRUE or FALSE
+#' @param rho the correlation within the covariates (default: 0)
+#' @param n_rules the number of causal rules, either 2 (default) or 4
+#' @param effect_size the effect size magnitude (default: 0.5)
+#' TODO: what is the range of effect size magnitude?
+#' @param binary whether to use binary or continuous outcomes,
+#'  either TRUE (default) or FALSE
 #'
 #' @return
 #' a list of synthetic data containing an outcome vector, a treatment vector,
@@ -22,7 +24,8 @@
 #'
 #' @export
 #'
-generate_cre_dataset <- function(n, rho, n_rules, effect_size, binary) {
+generate_cre_dataset <- function(n, rho = 0, n_rules = 2,
+                                 effect_size = 0.5, binary = TRUE) {
 
   # Check for correct binary input
   if (!(binary %in% c(TRUE, FALSE))) {