+++ NEWS.md
@@ -1,8 +1,5 @@
 ## CRE (devloping version) 
 
-### Added
-
-
 ### Changed
 - `estimate_cate` include two methods for estimating the CATE values
 - `cre` added initial checks for binary outcome and whether to include the propensity score in the ITE estimation
@@ -11,6 +8,7 @@
 - Moved `take1` function into `utils.R` 
 
 ### Added
+- `check_input_data` function
 - example to `generate_cre_dataset`
 - `generate_cre_dataset` function to generate synthetic data for testing the package
 - `test-generate_cre_dataset` function test